safari-534.53-branch/LayoutTests/fast/media/media-query-list-02-expected.txt

Test the MediaQueryList interface: http://dev.w3.org/csswg/cssom-view/#the-mediaquerylist-interface.

Testing addListener in different MediaQueryList objects and the callback call sequence

------------- minimized -------------
[1] - query (-webkit-view-mode: windowed) changed to false
[1] - query (-webkit-view-mode: minimized) changed to true
[2] - query (-webkit-view-mode: windowed) changed to false
[2] - query (-webkit-view-mode: minimized) changed to true
------------- maximized -------------
[1] - query (-webkit-view-mode: minimized) changed to false
[1] - query (-webkit-view-mode: maximized) changed to true
[2] - query (-webkit-view-mode: minimized) changed to false
[2] - query (-webkit-view-mode: maximized) changed to true
------------- windowed -------------
[1] - query (-webkit-view-mode: windowed) changed to true
[1] - query (-webkit-view-mode: maximized) changed to false
[2] - query (-webkit-view-mode: windowed) changed to true
[2] - query (-webkit-view-mode: maximized) changed to false
------------- floating -------------
[1] - query (-webkit-view-mode: windowed) changed to false
[1] - query (-webkit-view-mode: floating) changed to true
[2] - query (-webkit-view-mode: windowed) changed to false
[2] - query (-webkit-view-mode: floating) changed to true
